[Update] 修改Luna结构
