+++ package.json
@@ -19,6 +19,7 @@
     "@angular/core": "^5.2.11",
     "@angular/flex-layout": "^5.0.0-beta.15",
     "@angular/forms": "^5.2.11",
+    "@angular/http": "^5.2.11",
     "@angular/material": "^5.2.0",
     "@angular/platform-browser": "^5.2.11",
     "@angular/platform-browser-dynamic": "^5.2.11",
@@ -31,6 +32,7 @@
     "ajv": "^6.5.0",
     "animate.css": "^3.6.1",
     "body-parser": "^1.18.2",
+    "bootstrap": "^4.3.1",
     "clipboard": "^1.7.1",
     "compass-mixins": "^0.12.10",
     "core-js": "2.5.3",
@@ -75,7 +77,7 @@
   "devDependencies": {
     "@angular-devkit/core": "^0.4.9",
     "@angular-devkit/schematics": "^0.4.9",
-    "@angular/cli": "^7.2.1",
+    "@angular/cli": "^1.7.4",
     "@angular/compiler-cli": "^5.2.11",
     "@angular/language-service": "5.2.0",
     "@types/jasmine": "2.8.4",
