+++ src/app/elements/dialog/dialog.service.ts
@@ -1,6 +1,6 @@
 import {Component, Inject, Injectable, OnInit} from '@angular/core';
 import {MAT_DIALOG_DATA, MatDialog, MatDialogRef} from '@angular/material';
-import {LogService} from '../../app.service';
+import {LogService} from '@app/app.service';
 import {FormControl, Validators} from '@angular/forms';
 
 // import * as layer from 'layui-layer/src/layer.js';
