+++ src/app/elements/guacamole/guacamole.component.scss
@@ -2,3 +2,7 @@ iframe {
   border: none;
   background-color: white;
 }
+
+.rdpIframe {
+ height: 100%;
+}
