+++ src/app/elements/left-bar/left-bar.component.html
@@ -1,8 +1,8 @@
-<div class="sidebar" fxLayout="column" ngxSplit="column">
+<div class="sidebar" fxLayout="column" >
   <div fxflex="1 1 30px" class="tree-filter">
     <elements-tree-filter ></elements-tree-filter>
   </div>
-  <div class="overflow ngx-scroll-overlay" fxflex="1 1 calc(90%-60px)">
+  <div class="overflow ngx-scroll-overlay" fxflex="1 1 calc(100% - 60px)">
     <elements-asset-tree ></elements-asset-tree>
   </div>
