+++ src/app/elements/left-bar/left-bar.component.scss
@@ -3,10 +3,6 @@
   width: 100%;
   overflow: auto;
 }
-//
-//:root {
-//  font-family: "Hiragino Kaku Gothic ProN", Meiryo, sans-serif;
-//}
 
 label {
   margin-bottom: 0;
@@ -16,14 +12,10 @@ label {
   content: "\f26c";
 }
 
-.tree-filter {
-  height: 30px;
-}
-
 .overflow {
-  height: 90%;
   display: flex;
   width: 100%;
+  height: 100%;
   float: left;
   position: inherit;
   background: #2f2a2a;
@@ -38,7 +30,7 @@ label {
   padding: 1px 20px 0 20px;
   border-top: 1px solid #e7eaec;
   bottom: 0;
-  height: 60px;
+  height: 30px;
 }
 
 .footer-version > p {
@@ -49,23 +41,6 @@ label {
   color: #d6cbcb;
 }
 
-//@import "~@swimlane/ngx-ui/release/styles/components/scrollbars";
-
 .ngx-scroll-overlay {
   overflow: auto; // for FF
 }
-
-//.sidebar::-webkit-scrollbar-track {
-//  -webkit-box-shadow: inset 0 0 2px rgba(0, 0, 0, 0.3);
-//  background-color: #676a6c;
-//}
-//
-//.sidebar::-webkit-scrollbar {
-//  width: 8px;
-//}
-//
-//.sidebar::-webkit-scrollbar-thumb {
-//  background-color: #F5F5F5;
-//  border-radius: 6px;
-//  border: 2px solid transparent;
-//}
