+++ src/app/elements/nav/nav.component.html
@@ -1,9 +1,9 @@
-<script src="../../trans.pipe.spec.ts"></script>
 <div class="nav">
-  <ul>
-    <li><a href="/"><img src="static/imgs/logo.png" height="26px"/></a>
+  <ul class="nav-main">
+    <li>
+      <a href="/"><img src="static/imgs/logo.png" height="26px"/></a>
     </li>
-    <li *ngFor="let v of navs" [ngClass]="{'dropdown': v.children}">
+    <li *ngFor="let v of navs" [ngClass]="{'dropdown': v.children}" >
       <a>{{v.name|trans}}</a>
       <ul [ngClass]="{'dropdown-content': v.children}">
         <li *ngFor="let vv of v.children" [ngClass]="{'disabled': vv.disable}">
