--- src/app/elements/ofooter/ofooter.component.html
@@ -1,9 +1,0 @@
-<div class="footer fixed">
-    <div class="pull-right">
-        Version <strong>{{version}}</strong> GPLv2.
-        <img style="display: none" src="http://www.jumpserver.org/img/evaluate_avatar1.jpg">
-    </div>
-    <div>
-        <strong>Copyright</strong> 北京堆栈科技有限公司 &copy; 2014-2018
-    </div>
-</div>
