+++ src/app/elements/ssh-term/ssh-term.component.ts
@@ -1,9 +1,9 @@
 import {Component, Input, OnInit, OnDestroy } from '@angular/core';
 import {Terminal} from 'xterm';
-import {View} from '../content/model';
-import {LogService, UUIDService} from '../../app.service';
-import {Socket} from '../../utils/socket';
-import {getWsSocket, translate} from '../../globals';
+import {View} from '@app/model';
+import {LogService, UUIDService} from '@app/app.service';
+import {Socket} from '@app/utils/socket';
+import {getWsSocket, translate} from '@app/globals';
 
 
 @Component({
