+++ src/app/elements/tree-filter/tree-filter.component.ts
@@ -2,15 +2,31 @@ import {Component, OnInit, Pipe, PipeTransform} from '@angular/core';
 import {FormControl} from '@angular/forms';
 import {debounceTime, distinctUntilChanged} from 'rxjs/operators';
 
-import {LogService, TreeFilterService} from '../../app.service';
+import {LogService, TreeFilterService} from '@app/app.service';
+
+
+@Pipe({name: 'SearchFilter'})
+export class SearchFilter implements PipeTransform {
+  transform(value: any, input: string) {
+    if (input) {
+      input = input.toLowerCase();
+      return value.filter(function (el: any) {
+        // ToDo: search with a simple SQL like language, and a bug search a group's hosts
+        return JSON.stringify(el).toLowerCase().indexOf(input) > -1;
+      });
+    }
+    return value;
+  }
+}
 
 @Component({
   selector: 'elements-tree-filter',
   templateUrl: './tree-filter.component.html',
   styleUrls: ['./tree-filter.component.css'],
+  providers: [SearchFilter],
 })
 export class ElementTreeFilterComponent implements OnInit {
-  private searchControl: FormControl;
+  searchControl: FormControl;
   private debounce = 400;
 
   constructor(private _treeFilterService: TreeFilterService,
@@ -29,16 +45,3 @@ export class ElementTreeFilterComponent implements OnInit {
 }
 
 
-@Pipe({name: 'SearchFilter'})
-export class SearchFilter implements PipeTransform {
-  transform(value: any, input: string) {
-    if (input) {
-      input = input.toLowerCase();
-      return value.filter(function (el: any) {
-        // ToDo: search with a simple SQL like language, and a bug search a group's hosts
-        return JSON.stringify(el).toLowerCase().indexOf(input) > -1;
-      });
-    }
-    return value;
-  }
-}
