+++ src/app/pages/main/main.component.html
@@ -1,5 +1,5 @@
 <elements-nav></elements-nav>
-<div id="container" class="container-fluid row" fxLayout="row" ngxSplit="row">
+<div fxLayout="row" id="container" class="container-fluid row" ngxSplit="row">
   <div fxFlex="1 1 20%" minBasis="100px" maxBasis="800px" fxFlexFill ngxSplitArea *ngIf="DataStore.showLeftBar">
     <elements-left-bar></elements-left-bar>
   </div>
