+++ src/app/pages/not-found/not-found.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit} from '@angular/core';
-import {DataStore} from '../../globals';
+import {DataStore} from '@app/globals';
 
 @Component({
   selector: 'pages-not-found',
