--- src/assets/123.json
@@ -1,3 +1,0 @@
-{
-  "hello": "world"
-}
