+++ tsconfig.json
@@ -1,6 +1,11 @@
 {
   "compileOnSave": false,
   "compilerOptions": {
+    "baseUrl": "src",
+    "paths": {
+      "@src/*": ["*"],
+      "@app/*": ["app/*"]
+    },
     "allowJs": true,
     "outDir": "./dist/out-tsc",
     "sourceMap": true,
